optimizations

+++ ADXL345.ino
@@ -8,41 +8,41 @@
 #include "ADXL345.h"
 
 //I2C addresses 
-static int AccelAddress = 0x53; // Write:0xA6  Read:0xA7
+static uint8_t AccelAddress = 0x53; // Write:0xA6  Read:0xA7
 
 //============================================
 // Accelerometer
 //============================================
 void InitADXL345(void) {
     Wire.beginTransmission(AccelAddress);
-    Wire.write(0x2D);  // power register
-    Wire.write(0x08);  // measurement mode
+    Wire.write((uint8_t)0x2D);  // power register
+    Wire.write((uint8_t)0x08);  // measurement mode
     Wire.endTransmission();
     delay(20);
     Wire.beginTransmission(AccelAddress);
-    Wire.write(0x31);  // Data format register
-    Wire.write(0x08);  // set to full resolution
+    Wire.write((uint8_t)0x31);  // Data format register
+    Wire.write((uint8_t)0x08);  // set to full resolution
     Wire.endTransmission();
     delay(20);	
     // Because our main loop runs at 50Hz we adjust the output data rate to 50Hz (25Hz bandwith)
     // Wire.beginTransmission(AccelAddress);
-    // Wire.write(0x2C);  // Rate
-    // Wire.write(0x09);  // set to 50Hz, normal operation
+    // Wire.write((uint8_t)0x2C);  // Rate
+    // Wire.write((uint8_t)0x09);  // set to 50Hz, normal operation
     // Wire.endTransmission();
 }
 
 
 // Reads x,y and z vAccelerometer registers
 void ReadADXL345(void) {
-    int i = 0;
+    int8_t i = 0;
     byte buff[6];
 
     Wire.beginTransmission(AccelAddress); 
-    Wire.write(0x32);        //sends address to read from
+    Wire.write((uint8_t)0x32);        //sends address to read from
     Wire.endTransmission(); //end transmission
 
     Wire.beginTransmission(AccelAddress); //start transmission to device
-    Wire.requestFrom(AccelAddress, 6);    // request 6 bytes from device
+    Wire.requestFrom(AccelAddress, (uint8_t)6);    // request 6 bytes from device
 
     while (Wire.available()) {  // ((Wire.available())&&(i<6))
         buff[i] = Wire.read();  // receive one byte
